requirements.txt
-f /usr/share/pip-wheels
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.3
Flask==1.0.2
Flask-Caching==1.4.0
Flask-Login==0.4.1
Flask-Misaka==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.0
misaka==2.1.1
more-itertools==5.0.0
mysqlclient==1.3.14
oauthlib==2.1.0
passlib==1.7.1
pluggy==0.8.0
py==1.7.0
pycparser==2.19
PyJWT==1.7.1
pytest==4.0.2
pytest-cov==2.6.1
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
SQLAlchemy==1.3.3
SQLAlchemy-Utils==0.33.9
stripe==2.16.0
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
